Updated .Jar file
This one has an Apache Ant build file script with a link to the example build file via the Apache Ant website under view. Also, there is a description of Ant (My description, so it might not be very good). Under help/installed extensions is the purpose of the extension and also a link to the Apache Ant download.